+++ library/src/methods/index.ts
@@ -1,3 +1,4 @@
+export * from './brand/index.ts';
 export * from './coerce/index.ts';
 export * from './is/index.ts';
 export * from './keyof/index.ts';